refactor(storage): improve query normalization and split MCP tests
- Refactor Turso query cache normalization to strip block comments and use regex
- Add once_cell dependency for MSRV compatibility with Lazy
- Split large persistent_storage_tests.rs in do-memory-mcp into smaller files
- Add regression tests for multiline and commented SQL in dependency detection
