+++ memory-storage-turso/Cargo.toml
@@ -60,6 +60,10 @@ chrono = { workspace = true }
 # Logging
 tracing = { workspace = true }
 
+# Regular expressions for SQL parsing
+regex = { workspace = true }
+once_cell = { workspace = true }
+
 # Synchronization
 parking_lot = "0.12.5"
