+++ plans/GOAP_PR_532_RESOLUTION.md
@@ -12,34 +12,18 @@ Resolve all failing GitHub Actions and address all reviewer comments in PR 532 (
 ## Tasks
 
 ### Phase 1: Infrastructure & Infrastructure-adjacent Fixes
-- **T1.1: Fix `scripts/code-quality.sh`**
-  - Fix CLI parsing (remove unconditional `shift`, handle flags properly).
-  - Update error handling to capture exit codes from `cargo fmt`, `clippy`, and `audit` even if they fail.
-- **T1.2: Update `Cargo.toml`**
-  - Add dependency debug override in `[profile.dev.package."*"]`.
-  - Set `debug = "line-tables-only"` for overall dev profile.
-- **T1.3: Fix Markdown Linting (MD022)**
-  - Ensure headings are surrounded by blank lines in:
-    - `plans/adr/ADR-054-CloudEvents-Integration.md`
-    - `plans/GOAP_CLOUDEVENTS_INTEGRATION.md`
-    - `plans/GOAP_CLOUDEVENTS_STABILIZATION.md`
-- **T1.4: Fix `memory-core/src/retrieval/cascade/mod.rs`**
-  - Set `api_calls: 1` in the Tier 4 fallback branch.
+- [x] **T1.1: Fix `scripts/code-quality.sh`**
+- [x] **T1.2: Update `Cargo.toml`**
+- [x] **T1.3: Fix Markdown Linting (MD022)**
+- [x] **T1.4: Fix `memory-core/src/retrieval/cascade/mod.rs`**
 
 ### Phase 2: Storage Backend Refinements
-- **T2.1: Refactor `memory-storage-duckdb/src/storage/patterns.rs`**
-  - Sanitize raw JSON in `Error::Storage` messages (don't leak sensitive data).
-  - Handle `i64::try_from` error explicitly for `sample_size`.
-  - Use `match &pattern` to avoid moving bindings.
-- **T2.2: Refactor `memory-storage-turso/src/cache/query_cache_types.rs`**
-  - Update `from_query` to strip SQL comments and normalize whitespace before dependency detection.
+- [x] **T2.1: Refactor `memory-storage-duckdb/src/storage/patterns.rs`**
+- [x] **T2.2: Refactor `memory-storage-turso/src/cache/query_cache_types.rs`**
 
 ### Phase 3: Test Suite Optimization
-- **T3.1: Split `memory-mcp/tests/persistent_storage_tests.rs`**
-  - Break the file down into smaller, focused test files (≤ 500 lines).
-  - Create a shared helper for setup if needed.
-- **T3.2: Update `memory-storage-turso/src/cache/query_cache_tests.rs`**
-  - Add regression tests for multiline and commented SQL in dependency detection.
+- [x] **T3.1: Split `memory-mcp/tests/persistent_storage_tests.rs`**
+- [x] **T3.2: Update `memory-storage-turso/src/cache/query_cache_tests.rs`**
 
 ### Phase 4: Validation
 - **T4.1: Run Formatting & Linting**
